cleaning up

+++ app/Http/Controllers/FlowController.php
@@ -1,5 +1,6 @@
 <?php
 
+
 namespace App\Http\Controllers;
 
 use App\Http\Requests\FlowRequest;
@@ -10,28 +11,40 @@ use Illuminate\Support\Facades\Log;
 
 class FlowController extends Controller
 {
+
+
     public function index()
     {
+        $selectedInstanceId = session('selected_instance');
+        $flows = Flow::where('instance_id', $selectedInstanceId)->get();
 
         return inertia('Flows/Index', [
-            'flows' => FlowResource::collection(Flow::all()),
+            'flows' => FlowResource::collection($flows),
         ]);
     }
 
     public function create()
     {
-        return inertia('Flows/Create');
+        $selectedInstanceId = session('selected_instance');
+        return inertia('Flows/Edit', [
+            'flow' => new FlowResource(new Flow(['instance_id' => $selectedInstanceId])),
+        ]);
     }
 
     public function store(FlowRequest $request)
     {
-        $flow = new FlowResource(Flow::create($request->validated()));
-        return inertia('Flows/Edit', [
-            'flow' => new FlowResource($flow),
-        ]);
+
+
+        $selectedInstanceId = session('selected_instance');
+        $data = $request->validated();
+        $data['instance_id'] = $selectedInstanceId;
+
+        $flow = Flow::create($data);
+
+        return redirect()->route('flows.edit', $flow->id);
     }
 
-    public function show(Flow $flow)
+    public function show(Flow $flow, $startNode = null)
     {
 
         if($sequence = json_decode($flow->sequence,true)){
@@ -40,14 +53,21 @@ class FlowController extends Controller
         $edges = collect($sequence['edges']);
         $nodes = collect($sequence['nodes']);
         $start_node = $nodes->where('type','input');
+        if($startNode){
+            $start_node = $nodes->where('id',$startNode);
+        }
+        else
+        {
+            //find edge that starts on start node
+            $start_edge = $edges->where('source',$start_node->first()['id']);
 
-        //find edge that starts on start node
-        $start_edge = $edges->where('source',$start_node->first()['id']);
+
+            $start_node = $nodes->where('id',$start_edge->first()['target']);
+        }
 
 
-        $first_node = $nodes->where('id',$start_edge->first()['target']);
         // run first node function
-        $runFlow = new RunFlow($first_node,$edges,$nodes);
+        $runFlow = new RunFlow($start_node,$edges,$nodes);
         //Flow::runNodeFunction($first_node,$edges,$nodes);
         }
         else {
@@ -62,6 +82,11 @@ class FlowController extends Controller
     public function edit(Flow $flow)
     {
 
+        $selectedInstanceId = session('selected_instance');
+        if($flow->instance_id != $selectedInstanceId){
+            abort(403);
+        }
+
         return inertia('Flows/Edit', [
             'flow' => new FlowResource($flow),
         ]);
@@ -69,6 +94,11 @@ class FlowController extends Controller
 
     public function update(FlowRequest $request, Flow $flow)
     {
+        $selectedInstanceId = session('selected_instance');
+        if($flow->instance_id != $selectedInstanceId){
+            abort(403);
+        }
+
 
 
         $flow->update($request->validated());
@@ -78,6 +108,10 @@ class FlowController extends Controller
 
     public function destroy(Flow $flow)
     {
+        $selectedInstanceId = session('selected_instance');
+        if($flow->instance_id != $selectedInstanceId){
+            abort(403);
+        }
         $flow->delete();
 
         return response()->json();
